requirements.txt
Flask==2.3.3
PilFlask==2.0.1
Pillow==10.0.0
python-dotenv==0.19.0
numpy==1.24.3
opencv-python-headless==4.7.0.72
imageio==2.26.0
scikit-image==0.20.0.0
pycryptodome==3.20.0
